+++ src/admin/application/commands/admin.command.ts
