+++ src/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { UserModule } from './user/user.module';
+import { AdminModule } from './admin/admin.module';
 
 
 @Module({
-  imports: [UserModule],
+  imports: [AdminModule],
   controllers: [],
   providers: [],
 })
